WIP: Update `@percy/agent` for testing
This also adds `babel-present-env` which is defined in the webpack
config but not installed in the
project. `@cypress/webpack-preprocessor` does not have it as a
dependency, it's listed as an optional one. Surprised that didn't fail
sooner?

This also removes the `node: { fs: 'empty' }` workaround that was in
place. All tests pass!