Rd/add percy (#6)
* WIP: Update `@percy/agent` for testing

This also adds `babel-present-env` which is defined in the webpack
config but not installed in the
project. `@cypress/webpack-preprocessor` does not have it as a
dependency, it's listed as an optional one. Surprised that didn't fail
sooner?

This also removes the `node: { fs: 'empty' }` workaround that was in
place. All tests pass!

* Percy tests pass with change to `healthcheck`

* Update to latest released percy packages

This does not properly solve the winston / fs issue yet. That's coming
in a future release.